# documentation/swagger.yaml
openapi: 2.0
info:
  title: ical-relay
  description: |-
    ...
  contact:
    email: mail@julian-lemmerich.de
  license:
    name:  GNU GENERAL PUBLIC LICENSE
    url: https://www.gnu.org/licenses/gpl-3.0.html
  version: 2.0.0-beta
...
paths:
...
  /api/profiles/{profile}/checkAuth:
    get:
      tags:
        - admin
      summary: Checks Token Autorization
      description: Checks if token is authorized to access the profile.
      operationId: checkAuth
      parameters:
        - name: profile
          in: path
          description: Name of Profile to check Auth for.
          required: true
          schema:
            type: string
      security:
        - tokenAuth: []
      responses:
        '200':
          description: Authentication successful
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Profile not found
  /api/checkSuperAuth:
    get:
      tags:
        - admin
      summary: Checks Token Autorization
      description: Checks if token is authorized to access the profile.
      operationId: checkSuperAuth
      security:
        - tokenAuth: []
      responses:
        '200':
          description: Authentication successful
        '401':
          $ref: "#/components/responses/UnauthorizedError"
  /api/profiles/{profile}/rules:
    get:
      tags:
        - admin
      summary: Get all Rules of a Profile
      description: Get all Rules of a Profile
      operationId: getRules
      parameters:
        - name: profile
          in: path
          description: Name of Profile to get Rules for.
          required: true
          schema:
            type: string
      security:
        - tokenAuth: []
      responses:
        '200':
          $ref: "#/components/responses/RuleList"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Profile not found
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - admin
      summary: Add a Rule to a Profile
      description: General add a Rule to a Profile. token Auth allows access to the Modules edit-bysummary-regex and delete-bysummary-regex. Super Auth allows access to all Modules.
      operationId: addRule
      parameters:
        - name: profile
          in: path
          description: Name of Profile to add Rule to.
          required: true
          schema:
            type: string
        - name: rule
          in: body
          description: Rule to add
          required: true
          schema:
            $ref: "#/components/schemas/Rule"
      security:
        - tokenAuth: []
      responses:
        '200':
          $ref: "#/components/responses/RuleList"
        '400':
          description: Module not found
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Not allowed to add this Module. (Token Auth)
        '404':
          description: Profile not found
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token Authentication for specific profiles. Not all features availible.
    superAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token Authentication for all profiles and Server features. All features availible.
  responses:
    UnauthorizedError:
      description: Token is missing or invalid
      headers:
        Authorization:
          schema:
            type: string
            description: Token to use for authentication
    InternalError:
      description: Server Error
      content:
        plain/text:
          schema:
            type: string
            description: Error message
    RuleList:
      description: successful operation
      content:
        application/json:
          schema:
            example:
              - name: "delete-bysummary-regex"
                regex: "testentry"
                from: "2021-12-02T00:00:00Z"
                until: "2021-12-31T00:00:00Z"
              - name: "add-url"
                url: "https://othersource.com/othercalendar.ics"
                header-Cookie: "MY_AUTH_COOKIE=abcdefgh"
...